pyproject.toml
[project]
name = "proxy-outlook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "selenium>=4.34.2",
    "undetected-chromedriver>=3.5.5",
    "python-dotenv>=1.0.1",
    "setuptools>=70.0.0",
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "pyjwt>=2.8.0",
]
